+++ arguebuf/cli/model.py
@@ -1,5 +1,3 @@
-from __future__ import annotations
-
 import typing as t
 from dataclasses import dataclass
 from pathlib import Path
@@ -19,7 +17,7 @@ class PathPair:
         path_out: Path,
         input_glob: t.Optional[str],
         output_suffix: t.Optional[str],
-    ) -> t.List[PathPair]:
+    ) -> list["PathPair"]:
         pairs: list[PathPair] = []
 
         if path_in.is_file():
@@ -46,7 +44,8 @@ class PathPair:
                 files_out.append(file_out)
 
             pairs.extend(
-                cls(file_in, file_out) for file_in, file_out in zip(files_in, files_out)
+                cls(file_in, file_out)
+                for file_in, file_out in zip(files_in, files_out, strict=True)
             )
 
         else:
@@ -58,7 +57,7 @@ class PathPair:
         return pairs
 
     @staticmethod
-    def label(path_pair: t.Optional[PathPair]) -> str:
+    def label(path_pair: t.Optional["PathPair"]) -> str:
         """Generate a string for representing a path pair.
 
         Args: